Replace Arduino Modbus library with an own RTU server on the IDF UART driver
The bridge only ever answers two function codes on a fixed set of
registers, so the emelianov/modbus-esp8266 dependency bought little and
tied the component to the Arduino framework via ModbusRTU, Stream and
Serial2.

modbus_rtu.{h,cpp} implements just that: CRC16, address filtering and
frame handling. Frame boundaries come from the IDF UART driver's RX
timeout, which is what Modbus RTU's 3.5-character gap actually is -
more reliable than polling for it in software.

hoermann.cpp keeps its logic; the register map is now three plain
arrays and the former onSet callbacks take (old, new) values. Order is
preserved: prepare the response registers, apply the master's writes,
then answer.

No Arduino headers remain, so the component can build under esp-idf.

+++ components/hcpbridge/hcpbridge.cpp
@@ -1,5 +1,7 @@
 #include "hcpbridge.h"
 
+#include <cinttypes>
+
 namespace esphome {
 namespace hcpbridge {
 
@@ -17,7 +19,7 @@ void HCPBridge::add_on_state_callback(std::function<void()> &&callback, const ch
     auto start = millis();
     callback();
     auto end = millis();
-    ESP_LOGD(TAG, "Callback executed in %u ms [Tag: %s]", end - start, tag);
+    ESP_LOGD(TAG, "Callback executed in %" PRIu32 " ms [Tag: %s]", (uint32_t)(end - start), tag);
   };
   this->state_callback_.add(std::move(wrapped_callback));
 }
